Remove LINK_FLAGS and COMPILE_FLAGS properties from App-ompi
target_link_libraries(App-ompi PUBLIC MPI::MPI_C) does the job

`make VERBOSE=1` shows that the flags are unused when linking
libApp-ompi.a because we are creating an archive.

Changing the test to link with `App-ompi` and looking at its
linking command with `make VERBOSE=1` shows that because of the
public dependency on `MPI::MPI_C`, this executable receives
all the link flags that it needs.